pre-cleanup commit of refactored timing decorators. tested for maxwell conv, not done for cudanet

+++ neon/datasets/i1k.py
@@ -33,7 +33,7 @@ macroq_flag = False
 miniq_flag = False
 gpuq_flag = False
 
-BDTYPE = 'float32'
+BDTYPE = 'float16'
 logger.warning("i1k with hardcoded dtype %s" ,BDTYPE)
 
 def my_pickle(filename, data):
